do not initialize io-uring object if async scorer is not enabled
