+++ lib/segment/src/vector_storage/common.rs
@@ -1,0 +1,11 @@
+use std::sync::atomic::{AtomicBool, Ordering};
+
+static ASYNC_SCORER: AtomicBool = AtomicBool::new(false);
+
+pub fn set_async_scorer(async_scorer: bool) {
+    ASYNC_SCORER.store(async_scorer, Ordering::Relaxed);
+}
+
+pub fn get_async_scorer() -> bool {
+    ASYNC_SCORER.load(Ordering::Relaxed)
+}
