+++ lib/segment/src/vector_storage/raw_scorer.rs
@@ -9,6 +9,7 @@ use crate::spaces::metric::Metric;
 use crate::spaces::simple::{CosineMetric, DotProductMetric, EuclidMetric};
 use crate::spaces::tools::peek_top_largest_iterable;
 use crate::types::{Distance, PointOffsetType, ScoreType};
+use crate::vector_storage;
 
 /// Optimized scorer for multiple scoring requests comparing with a single query
 /// Holds current query and params, receives only subset of points to score
@@ -65,16 +66,6 @@ pub struct RawScorerImpl<'a, TMetric: Metric, TVectorStorage: VectorStorage> {
     pub is_stopped: &'a AtomicBool,
 }
 
-static ASYNC_SCORER: AtomicBool = AtomicBool::new(false);
-
-pub fn set_async_scorer(async_scorer: bool) {
-    ASYNC_SCORER.store(async_scorer, Ordering::Relaxed);
-}
-
-pub fn get_async_scorer() -> bool {
-    ASYNC_SCORER.load(Ordering::Relaxed)
-}
-
 pub fn new_stoppable_raw_scorer<'a>(
     vector: Vec<VectorElementType>,
     vector_storage: &'a VectorStorageEnum,
@@ -85,7 +76,7 @@ pub fn new_stoppable_raw_scorer<'a>(
         VectorStorageEnum::Simple(vs) => raw_scorer_impl(vector, vs, point_deleted, is_stopped),
 
         VectorStorageEnum::Memmap(vs) => {
-            if get_async_scorer() {
+            if vector_storage::common::get_async_scorer() {
                 #[cfg(target_os = "linux")]
                 match super::async_raw_scorer::new(vector.clone(), vs, point_deleted, is_stopped) {
                     Ok(raw_scorer) => return raw_scorer,
